Agrego test a cuenta

+++ src/main/java/domain/transporte/FedEx.java
@@ -3,7 +3,7 @@ package domain.transporte;
 public class FedEx implements StrategyEnvio{
 
     @Override
-    public Integer calcularCosto(TipoDeEnvio tipoDeEnvio,Paquete paquete) {
+    public double calcularCosto(TipoDeEnvio tipoDeEnvio, Paquete paquete) {
         return paquete.getPeso();
     }
