+++ src/main/java/domain/transporte/UPS.java
@@ -3,7 +3,7 @@ package domain.transporte;
 public class UPS implements StrategyEnvio{
 
     @Override
-    public Integer calcularCosto(TipoDeEnvio tipoDeEnvio,Paquete paquete) {
+    public double calcularCosto(TipoDeEnvio tipoDeEnvio, Paquete paquete) {
         return paquete.getPeso();
     }
